dagster/pyproject.toml
[project]
name = "finpy"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.29",
    "dagster>=1.12.11",
    "dagster-webserver>=1.12.11",
    "grpcio>=1.76.0",
    "grpcio-status>=1.76.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyspark[connect]==3.5.5",
    "python-dotenv>=1.2.1",
    "trino>=0.336.0",
]

[tool.dagster]
module_name = "pipelines.definitions"
code_location_name = "finpy"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
